Update UrbanStructureMain.R
This script uses the functions defined in UrbanStructureFunctions.R to generate a visualization of the city of Cuenca. This includes reading and transforming continent data, determining and preparing the bounding box for analysis, retrieving street lines, water and island areas from OpenStreetMap, splitting the streets where they intersect, and processing coordinates.